scaling temporal window

+++ src/processing/vCornerRT/include/vCornerRTThread.h
@@ -111,7 +111,7 @@ private:
 public:
 
     vCornerThread(unsigned int height, unsigned int width, std::string name, bool strict, int qlen,
-                  int temporalsize, int windowRad, int sobelsize, double sigma, double thresh, int nthreads);
+                  double temporalsize, int windowRad, int sobelsize, double sigma, double thresh, int nthreads);
     bool threadInit();
     bool open(std::string portname);
     void onStop();
